Pass tests @1000 repetition level on phone and mac
Need to refactor connectable and subject to allow for two channels: control and data.
